features/requirements.txt
edge-tts
pygame
pyaudio
vosk
spotipy
pyautogui
keyboard
pyttsx3
google-generativeai
python-dotenv
textblob
screen-brightness-control
pycaw
win32gui
pillow